WILL-1481 / WILL-1490 - Adding ignore, delete and ignoring extensions ### Feature: Ignore a Not Found entry An entry can now be ignored, so it won't be displayed in the overview, unless specifically filtered.
### Feature: Delete a Not Found entry
An entry can now be deleted, for instance if a redirect is created and the entry is now no longer expected to give 404.

### Feature: Ignore extensions
Any extensions inputted in the settings page, will be ignored when a 404 entry is inserted. For instance, if `xhtml` is added as an ignored extension, all URL's ending in `.xhtml` will be ignored and not result in an entry in the Not Found module.

### Feature: Create redirect
All URL's in the Not Found overview, will now have a quicklink to create a redirect in the WP Bonnier Redirect plugin
The WP Bonnier Redirect plugin has been updated to ^4.9 to support this feature.